Remove z4h & cleanup

+++ java/install.sh
@@ -1,8 +1,7 @@
 #!/usr/bin/env bash
 
 # install java
-if ! [ -f /Library/Java/JavaVirtualMachines/openjdk.jdk ]
-then
+if ! [ -f /Library/Java/JavaVirtualMachines/openjdk.jdk] && [ -f /usr/local/opt/openjdk/libexec/openjdk.jdk ]; then
   echo "setting java to use brew install openjdk"
   sudo ln -sfn /usr/local/opt/openjdk/libexec/openjdk.jdk /Library/Java/JavaVirtualMachines/openjdk.jdk
 fi
